mlops/requirements.txt
black
flake8
isort
loguru
mkdocs
python-dotenv
tqdm
typer
timm
Pillow
matplotlib
numpy
requests
duckduckgo-search
torch
torchvision
torchaudio
-f https://download.pytorch.org/whl/cu124
jupyter
-e .
